+++ connect/common/models.py
@@ -22,12 +22,15 @@ from model_utils import FieldTracker
 
 from connect import billing
 from connect.api.v1.internal.flows.flows_rest_client import FlowsRESTClient
-from connect.api.v1.internal.intelligence.intelligence_rest_client import \
-    IntelligenceRESTClient
+from connect.api.v1.internal.intelligence.intelligence_rest_client import (
+    IntelligenceRESTClient,
+)
 from connect.authentication.models import User
 from connect.billing.gateways.stripe_gateway import StripeGateway
-from connect.common.exceptions import (OrganizationAuthorizationException,
-                                       ProjectAuthorizationException)
+from connect.common.exceptions import (
+    OrganizationAuthorizationException,
+    ProjectAuthorizationException,
+)
 from connect.common.gateways.rocket_gateway import Rocket
 from connect.common.helpers import send_mass_html_mail
 from connect.internals.event_driven.producer.rabbitmq_publisher import RabbitmqPublisher
@@ -299,11 +302,11 @@ class Organization(models.Model):
 
 
 class OrganizationLevelRole(Enum):
-    NOTHING, VIEWER, CONTRIBUTOR, ADMIN, FINANCIAL, SUPPORT = list(range(6))
+    NOTHING, VIEWER, CONTRIBUTOR, ADMIN, FINANCIAL, SUPPORT, MARKETING = list(range(7))
 
 
 class OrganizationRole(Enum):
-    NOT_SETTED, VIEWER, CONTRIBUTOR, ADMIN, FINANCIAL, SUPPORT = list(range(6))
+    NOT_SETTED, VIEWER, CONTRIBUTOR, ADMIN, FINANCIAL, SUPPORT, MARKETING = list(range(7))
 
 
 class OrganizationAuthorization(models.Model):
@@ -319,6 +322,7 @@ class OrganizationAuthorization(models.Model):
         (OrganizationRole.VIEWER.value, _("viewer")),
         (OrganizationRole.FINANCIAL.value, _("financial")),
         (OrganizationRole.SUPPORT.value, _("support")),
+        (OrganizationRole.MARKETING.value, _("marketing")),
     ]
 
     uuid = models.UUIDField(
@@ -357,6 +361,9 @@ class OrganizationAuthorization(models.Model):
         if self.role == OrganizationRole.SUPPORT.value:
             return OrganizationLevelRole.SUPPORT.value
 
+        if self.role == OrganizationRole.MARKETING.value:
+            return OrganizationLevelRole.MARKETING.value
+
     @property
     def can_read(self):
         return self.level in [
@@ -365,6 +372,7 @@ class OrganizationAuthorization(models.Model):
             OrganizationLevelRole.ADMIN.value,
             OrganizationLevelRole.VIEWER.value,
             OrganizationLevelRole.SUPPORT.value,
+            OrganizationLevelRole.MARKETING.value,
         ]
 
     @property
@@ -373,6 +381,7 @@ class OrganizationAuthorization(models.Model):
             OrganizationLevelRole.CONTRIBUTOR.value,
             OrganizationLevelRole.ADMIN.value,
             OrganizationLevelRole.SUPPORT.value,
+            OrganizationLevelRole.MARKETING.value,
         ]
 
     @property
@@ -380,6 +389,7 @@ class OrganizationAuthorization(models.Model):
         return self.level in [
             OrganizationLevelRole.ADMIN.value,
             OrganizationLevelRole.SUPPORT.value,
+            OrganizationLevelRole.MARKETING.value,
         ]
 
     @property
@@ -399,6 +409,7 @@ class OrganizationAuthorization(models.Model):
             OrganizationLevelRole.ADMIN.value,
             OrganizationLevelRole.FINANCIAL.value,
             OrganizationLevelRole.SUPPORT.value,
+            OrganizationLevelRole.MARKETING.value,
         ]
 
     @property
@@ -657,8 +668,9 @@ class Project(models.Model):
         return created, data
 
     def create_chats_project(self):
-        from connect.api.v1.internal.chats.chats_rest_client import \
-            ChatsRESTClient  # to avoid circular import
+        from connect.api.v1.internal.chats.chats_rest_client import (
+            ChatsRESTClient,
+        )  # to avoid circular import
 
         created = False
         chats_client = ChatsRESTClient()
@@ -743,8 +755,7 @@ class Project(models.Model):
 
     def get_contacts(self, before: str, after: str, counting_method: str = None):
         from connect.billing.models import Contact
-        from connect.billing.utils import (custom_get_attendances,
-                                           get_attendances)
+        from connect.billing.utils import custom_get_attendances, get_attendances
 
         if not counting_method:
             counting_method = self.organization.organization_billing.plan_method
@@ -875,11 +886,11 @@ class ChatsAuthorization(models.Model):
 
 
 class ProjectRole(Enum):
-    NOT_SETTED, VIEWER, CONTRIBUTOR, MODERATOR, SUPPORT, CHAT_USER = list(range(6))
+    NOT_SETTED, VIEWER, CONTRIBUTOR, MODERATOR, SUPPORT, CHAT_USER, MARKETING = list(range(7))
 
 
 class ProjectRoleLevel(Enum):
-    NOTHING, VIEWER, CONTRIBUTOR, MODERATOR, SUPPORT, CHAT_USER = list(range(6))
+    NOTHING, VIEWER, CONTRIBUTOR, MODERATOR, SUPPORT, CHAT_USER, MARKETING = list(range(7))
 
 
 class ProjectAuthorization(models.Model):
@@ -893,6 +904,7 @@ class ProjectAuthorization(models.Model):
         (ProjectRole.MODERATOR.value, _("moderator")),
         (ProjectRole.SUPPORT.value, _("support")),
         (ProjectRole.CHAT_USER.value, _("Chat user")),
+        (ProjectRole.MARKETING.value, _("marketing")),
     ]
     uuid = models.UUIDField(
         _("UUID"), primary_key=True, default=uuid4.uuid4, editable=False
@@ -930,12 +942,15 @@ class ProjectAuthorization(models.Model):
             return ProjectRoleLevel.VIEWER.value
         elif self.role == ProjectRole.SUPPORT.value:
             return ProjectRoleLevel.SUPPORT.value
+        elif self.role == ProjectRole.MARKETING.value:
+            return ProjectRoleLevel.MARKETING.value
 
     @property
     def is_moderator(self):
         return self.level in [
             ProjectRoleLevel.MODERATOR.value,
             ProjectRoleLevel.SUPPORT.value,
+            ProjectRoleLevel.MARKETING.value,
         ]
 
     @property
@@ -943,6 +958,7 @@ class ProjectAuthorization(models.Model):
         return self.level in [
             ProjectRoleLevel.MODERATOR.value,
             ProjectRoleLevel.SUPPORT.value,
+            ProjectRoleLevel.MARKETING.value,
         ]
 
     @property
@@ -953,6 +969,7 @@ class ProjectAuthorization(models.Model):
             ProjectRoleLevel.VIEWER.value,
             ProjectRoleLevel.SUPPORT.value,
             ProjectRoleLevel.CHAT_USER.value,
+            ProjectRoleLevel.MARKETING.value,
         ]
 
     @property
@@ -961,6 +978,7 @@ class ProjectAuthorization(models.Model):
             ProjectRoleLevel.MODERATOR.value,
             ProjectRoleLevel.CONTRIBUTOR.value,
             ProjectRoleLevel.SUPPORT.value,
+            ProjectRoleLevel.MARKETING.value,
         ]